+++ modules/generic/src/main/scala/pureconfig/generic/CoproductHint.scala
@@ -3,7 +3,9 @@ package pureconfig.generic
 import com.typesafe.config.{ ConfigFactory, ConfigObject, ConfigValue, ConfigValueType }
 import pureconfig._
 import pureconfig.error._
+import pureconfig.generic.error.{ NoValidCoproductChoiceFound, UnexpectedValueForFieldCoproductHint }
 import pureconfig.syntax._
+import shapeless.CNil
 
 /**
  * A trait that can be implemented to disambiguate between the different options of a coproduct or sealed family.
@@ -46,6 +48,16 @@ trait CoproductHint[T] {
    * @return `true` if the next class or coproduct option should be tried, `false` otherwise.
    */
   def tryNextOnFail(name: String): Boolean
+
+  /**
+   * Returns a failed `ConfigReader` result scoped into the context of a `ConfigCursor` at the sealed family option,
+   * representing the failure to disambiguate the config value.
+   *
+   * @param cur a `ConfigCursor` at the sealed family option
+   * @return a failed `ConfigReader` result scoped into the context of a `ConfigCursor`.
+   */
+  def failToDisambiguate(cur: ConfigCursor): ConfigReader.Result[CNil] =
+    cur.failed(NoValidCoproductChoiceFound(cur.value))
 }
 
 /**
@@ -54,10 +66,12 @@ trait CoproductHint[T] {
  * This hint will cause derived `ConfigConvert` instance to fail to convert configs to objects if the object has a
  * field with the same name as the disambiguation key.
  *
- * By default, the field value written is the class or coproduct option name converted to lower case. This mapping can
+ * By default, the field value written is the class or coproduct option name converted to kebab case. This mapping can
  * be changed by overriding the method `fieldValue` of this class.
  */
 class FieldCoproductHint[T](key: String) extends CoproductHint[T] {
+
+  private val fieldTransformation: String => String = ConfigFieldMapping(PascalCase, KebabCase)
 
   /**
    * Returns the field value for a class or coproduct option name.
@@ -65,7 +79,7 @@ class FieldCoproductHint[T](key: String) extends CoproductHint[T] {
    * @param name the name of the class or coproduct option
    * @return the field value associated with the given class or coproduct option name.
    */
-  protected def fieldValue(name: String): String = name.toLowerCase
+  protected def fieldValue(name: String): String = fieldTransformation(name)
 
   def from(cur: ConfigCursor, name: String): ConfigReader.Result[Option[ConfigCursor]] = {
     for {
@@ -90,6 +104,9 @@ class FieldCoproductHint[T](key: String) extends CoproductHint[T] {
   }
 
   def tryNextOnFail(name: String) = false
+
+  override def failToDisambiguate(cur: ConfigCursor): ConfigReader.Result[CNil] =
+    cur.fluent.at(key).cursor.flatMap(cur => cur.failed(UnexpectedValueForFieldCoproductHint(cur.value)))
 }
 
 /**