+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ### 0.10.3 (unreleased)
 
+- Breaking changes
+  - The default transformation in `FieldCoproductHint` changed from converting class names to lower case to converting
+    them to kebab case (e.g. `ConfigOption` is converted to `config-option`). The old behavior can be retained by
+    putting in scope an `implicit def coproductHint[T] = new FieldCoproductHint[T]("type") { override def
+    fieldValue(name: String): String = name.toLowerCase }`;
+
 - New features
   - Added `deriveEnumerationReader`, `deriveEnumerationWriter` and `deriveEnumerationConvert` to the
     `pureconfig.generic.semiauto` package, allowing the derivation of readers and writers for enumerations encoded as